# Fix ingester locking