**What this PR does**:
Fixes #2534 by keeping the lock until the search is complete, but also splits out a separate mutex for the headblock ease contention on the core write path (live traces -> disk).   I added a new `BenchmarkInstanceContention` to help measure the effect, which is similar to the `xxxDoesNotRace` methods but prints out the op/sec of each thing.   Comparison between `main` and this branch are below. There's definitely a draw back in reads due to the increased lock time, but the write path actually looks better, since it only has to lock the headblock, and I took care to ensure that the read paths release the headblock as soon as possible.

```
name                   old time/op            new time/op            delta
InstanceContention-12             2.37s ±21%             2.69s ±29%      ~     (p=0.105 n=10+10)

name                   old blockflushes/sec   new blockflushes/sec   delta
InstanceContention-12              257 ±124%              886 ±101%  +244.84%  (p=0.000 n=10+10)

name                   old finds/sec          new finds/sec          delta
InstanceContention-12             6.84k ±20%             0.10k ±78%   -98.58%  (p=0.000 n=10+10)

name                   old pushes/sec         new pushes/sec         delta
InstanceContention-12             2.83k ±16%             2.39k ±28%   -15.51%  (p=0.009 n=10+10)

name                   old retentions/sec     new retentions/sec     delta
InstanceContention-12             1.91k ±22%             4.82k ±72%  +151.49%  (p=0.000 n=10+10)

name                   old searchTags/sec     new searchTags/sec     delta
InstanceContention-12               612 ±19%                52 ±36%   -91.51%  (p=0.000 n=10+10)

name                   old searchedBytes/sec  new searchedBytes/sec  delta
InstanceContention-12             25.2M ±13%             29.9M ±37%   +18.80%  (p=0.017 n=9+10)

name                   old searches/sec       new searches/sec       delta
InstanceContention-12               191 ±21%                34 ±20%   -82.00%  (p=0.000 n=10+9)

name                   old traceflushes/sec   new traceflushes/sec   delta
InstanceContention-12             5.43 ±126%              9.83 ±30%   +81.00%  (p=0.006 n=9+10)
```

**Note** Having the two mutexes makes it possible to introduce deadlocks if they are accessed in the wrong order. Left a couple huge comments explaining it, but the brittleness of future dev error is something to consider.

**Note2** I have been debating between this change and an alternate approach [here](https://github.com/grafana/tempo/compare/main...mdisibio:locklist).  The alternative approach is the `granular per-block locking nirvana` that I would like to see.  It reduces contention even more but it's not quite enough to balance out the increased complexity yet.  I'll keep working on it if there are no objections.

**Which issue(s) this PR fixes**:
Fixes #2534 

**Checklist**
- [x] Tests updated
- [ ] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`